# lib: use 32-bit atomics, s/pow/mul