@eqvinox 2017-07-11
Well, it's hidden behind the cost of [q]malloc(). That said, the number of cycles spent in `hash_get` with stats is triple of that of `hash_get` without stats.

```
WITHOUT stats:

Samples: 24K of event 'cycles', Event count (approx.): 2245909157
  Children      Self  Command   Shared Object       Symbol
+   87.22%     7.96%  hashtest  hashtest            [.] main
+   86.09%     0.00%  hashtest  hashtest            [.] _start
+   86.09%     0.00%  hashtest  libc-2.23.so        [.] __libc_start_main
+   82.60%    17.17%  hashtest  libfrr.so.0.0.0     [.] hash_get
+   59.21%    14.41%  hashtest  libfrr.so.0.0.0     [.] qcalloc
+   44.50%    11.53%  hashtest  libc-2.23.so        [.] __libc_calloc
+   32.92%    20.08%  hashtest  libc-2.23.so        [.] _int_malloc

1.08s
385622602 cycles in hash_get

WITH stats (b81352f...):

Samples: 31K of event 'cycles', Event count (approx.): 2970268704
  Children      Self  Command   Shared Object      Symbol
+   90.66%    38.95%  hashtest  libfrr.so.0.0.0    [.] hash_get
+   89.76%     1.97%  hashtest  hashtest           [.] main
+   88.66%     0.00%  hashtest  hashtest           [.] _start
+   88.66%     0.00%  hashtest  libc-2.23.so       [.] __libc_start_main
+   46.75%    10.34%  hashtest  libfrr.so.0.0.0    [.] qcalloc
+   36.31%    10.60%  hashtest  libc-2.23.so       [.] __libc_calloc
+   25.63%    15.54%  hashtest  libc-2.23.so       [.] _int_malloc

1.38s
1156919660 cycles in hash_get

1.38 / 1.08 = 1.277...
1156919660. / 385622602. = 3.000134...
```

Now whether this matters is a separate question.

[mismatches in perf output are caused by stack-trace recording being nonreliable]