@qlyoung 2017-07-11
Great. I guess we are reduced to `size_t` then.