@donaldsharp 2017-07-12
@eqvinox do we still need your PR open for this same issue?  and does the new changes that @qlyoung put in meet your requirements?